We do a little refactoring
