+++ UI/components/AlgoliaAdapters.js
@@ -1,0 +1,7 @@
+import TableHitsAlgoliaAdapter from "./TableHitsAlgoliaAdapter";
+import PaginateAlgoliaAdapter from "./PaginateAlgoliaAdapter";
+
+export {
+    TableHitsAlgoliaAdapter as TableHitsAdapter,
+    PaginateAlgoliaAdapter as PaginateAdapter,
+};
